+++ test/test_RHS_matlab.py
@@ -1,7 +1,7 @@
 # %%
 import matplotlib.pyplot as plt
 import numpy as np
-import RHS_filters
+import RHS_implementation
 
 
 def test_filterbank(rhs_bank, matlab_bank):
@@ -20,7 +20,9 @@ def test_RHSconv_matlab(matlab_filteroutput, matlab_bank, stimulus):
     filters_output = np.empty(matlab_bank.shape)
     for i in range(matlab_bank.shape[0]):
         for j in range(matlab_bank.shape[1]):
-            filters_output[i, j, ...] = RHS_filters.ourconv(stimulus, matlab_bank[i, j, ...])
+            filters_output[i, j, ...] = RHS_implementation.ourconv(
+                stimulus, matlab_bank[i, j, ...]
+            )
 
     assert np.allclose(matlab_filteroutput, filters_output)
 
@@ -30,7 +32,7 @@ def test_RHSconv_RHS(matlab_filteroutput, stimulus, rhs_bank):
     filters_output = np.empty(rhs_bank.shape)
     for i in range(rhs_bank.shape[0]):
         for j in range(rhs_bank.shape[1]):
-            filters_output[i, j, ...] = RHS_filters.ourconv(stimulus, rhs_bank[i, j, ...])
+            filters_output[i, j, ...] = RHS_implementation.ourconv(stimulus, rhs_bank[i, j, ...])
 
     assert np.allclose(matlab_filteroutput, filters_output)
 
@@ -40,7 +42,7 @@ def test_ODOG(stimulus, rhs_bank, output_odog_matlab):
     filters_output = np.empty(rhs_bank.shape)
     for i in range(rhs_bank.shape[0]):
         for j in range(rhs_bank.shape[1]):
-            filters_output[i, j, ...] = RHS_filters.ourconv(stimulus, rhs_bank[i, j, ...])
+            filters_output[i, j, ...] = RHS_implementation.ourconv(stimulus, rhs_bank[i, j, ...])
 
-    output = RHS_filters.odog_normalize(filters_output)
+    output = RHS_implementation.odog_normalize(filters_output)
     assert np.allclose(output, output_odog_matlab)
