+++ test/test_apply.py
@@ -1,5 +1,5 @@
 import numpy as np
-import RHS_filters
+import RHS_implementation
 
 from multyscale import filters
 
@@ -30,7 +30,7 @@ def test_conv_apply(rhs_bank, stimulus):
     for i in range(rhs_bank.shape[0]):
         for j in range(rhs_bank.shape[1]):
             f = rhs_bank[i, j]
-            o_conv[i, j, ...] = RHS_filters.ourconv(stimulus, f)
+            o_conv[i, j, ...] = RHS_implementation.ourconv(stimulus, f)
             o_apply[i, j, ...] = filters.apply(stimulus, f, pad=True)
 
     assert np.allclose(o_conv, o_apply)
