+++ test/test_filters.py
@@ -1,12 +1,12 @@
 # %%
 import matplotlib.pyplot as plt
 import numpy as np
-import RHS_filters
+import RHS_implementation
 
 import multyscale
 
 # %% RHS bank
-rhs_bank = RHS_filters.filterbank()
+rhs_bank = RHS_implementation.filterbank()
 
 # %% Parameters of image
 shape = (1024, 1024)  # filtershape in pixels
@@ -26,7 +26,7 @@ def test_circular_Gaussian():
     sigmas = np.array([1, 1]) * sigma1
     f = multyscale.filters.gaussian2d(x, y, (sigmas[0], sigmas[1]))
     f = f / f.sum()
-    f_2 = RHS_filters.d2gauss(shape[0], sigmas[0] * 32, shape[1], sigmas[0] * 32, 0)
+    f_2 = RHS_implementation.d2gauss(shape[0], sigmas[0] * 32, shape[1], sigmas[0] * 32, 0)
 
     plt.subplot(2, 2, 1)
     plt.imshow(f)
@@ -48,7 +48,9 @@ def test_elliptical_Gaussian():
     sigmas = np.array([1, 1]) * np.array([sigma1, sigma2])
     f = multyscale.filters.gaussian2d(x, y, (sigmas[0], sigmas[1]), orientation=orientation)
     f = f / f.sum()
-    f_2 = RHS_filters.d2gauss(shape[0], sigmas[0] * 32, shape[1], sigmas[1] * 32, orientation)
+    f_2 = RHS_implementation.d2gauss(
+        shape[0], sigmas[0] * 32, shape[1], sigmas[1] * 32, orientation
+    )
 
     plt.subplot(2, 2, 1)
     plt.imshow(f)
@@ -67,7 +69,7 @@ def test_ODOG():
     orientation = 150
     sigma3 = 2
     sigmas = np.array([[1, 1], [1, 2]]) * sigma3
-    rhs_odog = RHS_filters.odog(shape[0], shape[1], sigma3 * 32, orientation=orientation)
+    rhs_odog = RHS_implementation.odog(shape[0], shape[1], sigma3 * 32, orientation=orientation)
     multy_odog = multyscale.filters.odog(x, y, sigmas, orientation=(orientation, orientation))
 
     plt.subplot(2, 2, 1)
